requirements.txt
discord.py
python-dotenv
Pillow
requests
aiohttp
psycopg2-binary
datetime
pytz
